Place scribbu types & functions in module `(scribbu)`, even when defined in C.
This commit will place the snarfed functions provided by
libscribbu.so into module `(scribbu)` and updates
`scribbu.scm` to load the extension.

This commit also changes the linking between `scribbu` and
`libscribbu`.

I have squashed a number of commits on master-- this is generally
not a good idea, since I'm re-writing git history on a published
branch. That said, this project is obscure enough that I very much
doubt I'm inconveniencing anyone.