Fix the "delete" function in `scribbu text`.
This patch will fix a bug in the `-d` flag to `scribbu text`,
which is supposed to delete text frames. Previously, although
the code would erase the given frame from all lookup tables,
it would neglect to actually delete the frame from the
vector of frames in each tag type. This commit fixes that.